update code to seperate out streamlit functions
